Change maxcount message to be printed to stderr instead for better piping support
Before this commit one had the problem that when you want to pipe unicode output
e.g. to the clipboard like so `unicode --format="{pchar}" -m1 gear 2> /dev/null | xclip`
the error message was printed to stdout instead of the stderr which made it impossible
to ignore the error and made the error message be part of the clipboard.

With this commit above command would only include the character.